conduit_resolver: make the framework file a link
Instead of the chef file for conduit being a link to the
crowbar_framework file, reverse it so when dealing with chef tools that
cant deal with links, those tools work
